Don't use unexpected shared DefaultCredentialsProvider
`DefaultCredentialsProvider.create()` has a misleading name and documentation
and actually returns a shared instance. If we use it, then it will be closed
as part of closing the SsmAsyncClient and the rest of the application will
not be able to use it for any credential source that requires an API call (e.g.
workloads in EKS would not be able to refresh credentials via STS).